Added rating to user model, added flters

+++ cleaning_service/final_project/core/models/reviews.py
@@ -9,7 +9,7 @@ class Review(models.Model):
     request = models.ForeignKey(to=Request, on_delete=models.CASCADE, null=False)
     customer = models.ForeignKey(to=User, on_delete=models.CASCADE, null=False)
     feedback = models.TextField(verbose_name="Customer`s feedback", null=False)
-    rate = models.PositiveIntegerField(verbose_name="Star-rating", default=0, null=False, validators=(
+    rate = models.PositiveSmallIntegerField(verbose_name="Star-rating", default=0, null=False, validators=(
         MaxValueValidator(5), MinValueValidator(0)))
     created_at = models.DateTimeField(verbose_name="Time of creation", default=datetime.now(), null=False)
